+++ clients/python/src/agenticdriver/__init__.py
@@ -176,8 +176,8 @@ __all__ += [
 from .models import SessionMode, SessionIdentity, SessionHandle, SessionCreate, SessionInfo, SessionSnapshot, SessionDeleteResult
 __all__ += ['SessionMode', 'SessionIdentity', 'SessionHandle', 'SessionCreate', 'SessionInfo', 'SessionSnapshot', 'SessionDeleteResult']
 
-from .management import ProviderConfiguration, ConfigureProvider, ManagementSnapshot
-__all__ += ["ProviderConfiguration", "ConfigureProvider", "ManagementSnapshot"]
+from .management import ProviderConfiguration, ConfigureProvider, ManagementSnapshot, ProviderDefinition, ProviderConnectionMethod
+__all__ += ["ProviderConfiguration", "ConfigureProvider", "ManagementSnapshot", "ProviderDefinition", "ProviderConnectionMethod"]
 
 from .connections import ConnectionGrant, CreateInvitation, ConnectionInfo, ConnectionInvitation, ConnectionCredentials, ConnectionList, connection_target
 __all__ += ["ConnectionGrant", "CreateInvitation", "ConnectionInfo", "ConnectionInvitation", "ConnectionCredentials", "ConnectionList", "connection_target"]